Jorgelo/adding camp role

+++ lib/Effects/Effects/HueSpin.h
@@ -1,9 +1,12 @@
 void hueSpinEffectLoop(struct EffectDataPacket *effectState, uint16_t loopMs)
 {
-  const uint8_t loopPosition = effectState->loopPosition % loopMs;
+
+  const uint16_t loopPosition = effectState->loopPosition % loopMs;
   const float loopPercent = float(loopPosition) / loopMs;
 
-  const int loopOffset = loopPercent * LED_CNT;
+  const uint16_t loopOffset = loopPercent * LED_CNT;
+
+  // TODO(jorglo): Since loopOffset is an int, it's a little jumpy;
 
   for (uint16_t i = 0; i < LED_CNT; i++)
   {
